assorted minor changes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4676 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ kdtree/include/pcl/kdtree/kdtree.h
@@ -340,10 +340,10 @@ namespace pcl
       }
 
       /** \brief Get the minimum allowed number of k nearest neighbors points that a viable result must contain. */
-      inline float
+      inline int
       getMinPts ()
       {
-        return ((float) min_pts_);
+        return (min_pts_);
       }
 
     protected:
